backend/requirements.txt
alembic==1.16.5
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.2.6
django-cors-headers==4.9.0
django-prometheus==2.4.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
fastapi==0.116.2
git-filter-repo==2.47.0
greenlet==3.2.4
h11==0.16.0
httptools==0.6.4
idna==3.10
kombu==5.5.4
Mako==1.3.10
MarkupSafe==3.0.2
msgpack==1.1.1
packaging==25.0
pillow==11.3.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
psycopg==3.2.10
psycopg2-binary==2.9.10
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
redis==6.4.0
reportlab==4.4.4
requests==2.32.5
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlparse==0.5.3
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
